+++ persistence/memory.go
@@ -1,4 +1,4 @@
-package persistance
+package persistence
 
 import "strings"
